+++ fastlane_core/lib/fastlane_core/ui/errors/fastlane_crash.rb
@@ -1,15 +1,17 @@
 module FastlaneCore
-  class FastlaneCrash < Interface::FastlaneException
-    def prefix
-      '[FASTLANE_CRASH]'
-    end
+  class Interface
+    class FastlaneCrash < FastlaneException
+      def prefix
+        '[FASTLANE_CRASH]'
+      end
 
-    def trimmed_backtrace
-      trim_backtrace(method_name: 'crash!')
-    end
+      def trimmed_backtrace
+        trim_backtrace(method_name: 'crash!')
+      end
 
-    def could_contain_pii?
-      caused_by_calling_ui_method?(method_name: 'crash!')
+      def could_contain_pii?
+        caused_by_calling_ui_method?(method_name: 'crash!')
+      end
     end
   end
 end
